tools/migration-unified-annotation-versioning/requirements.txt
grpcio
argparse
psycopg2
protobuf
rich
